OMAP4: PM: Fix prm/cm accessor api's for register access
OMAP4 has clockdomains spread across CM1/CM2 and PRM space, while power
domains are in PRM and MPU/CHIRON space. This makes it difficult to use
prm/cm accessor api's for register access and makes it necessary for
register absolute address to be stored for OMAP4, unlike register offsets
being used for OMAP3/2.

This patch tries to address the issue by defining all CM2 module offsets
from CM1 base, and all MPU/CHIRON module offsets from PRM base. Also the
2 CM modules in PRM (WKUP_CM and EMU_CM) have their offsets defined from
CM1 base. This makes it possible for anyone to use a prm/cm accessor api's
on OMAP4 seamlessly without knowing which address space the domains
physically fall under.

+++ arch/arm/mach-omap2/cm.h
@@ -106,20 +106,20 @@
 
 #ifndef __ASSEMBLER__
 
-extern u32 cm_read_mod_reg(s16 module, u16 idx);
-extern void cm_write_mod_reg(u32 val, s16 module, u16 idx);
-extern u32 cm_rmw_mod_reg_bits(u32 mask, u32 bits, s16 module, s16 idx);
+extern u32 cm_read_mod_reg(s32 module, u16 idx);
+extern void cm_write_mod_reg(u32 val, s32 module, u16 idx);
+extern u32 cm_rmw_mod_reg_bits(u32 mask, u32 bits, s32 module, s16 idx);
 
 extern int omap2_cm_wait_module_ready(s16 prcm_mod, u8 idlest_id,
 				      u8 idlest_shift);
 extern int omap4_cm_wait_module_ready(void __iomem *clkctrl_reg);
 
-static inline u32 cm_set_mod_reg_bits(u32 bits, s16 module, s16 idx)
+static inline u32 cm_set_mod_reg_bits(u32 bits, s32 module, s16 idx)
 {
 	return cm_rmw_mod_reg_bits(bits, bits, module, idx);
 }
 
-static inline u32 cm_clear_mod_reg_bits(u32 bits, s16 module, s16 idx)
+static inline u32 cm_clear_mod_reg_bits(u32 bits, s32 module, s16 idx)
 {
 	return cm_rmw_mod_reg_bits(bits, 0x0, module, idx);
 }
